+++ src/main/java/ru/practicum/shareit/item/comment/CommentDto.java
@@ -11,7 +11,6 @@ import java.time.LocalDateTime;
 public class CommentDto {
     private Long id;
     private String text;
-    private Item item;
     private String authorName;
     private LocalDateTime created;
 }
