+++ src/main/java/ru/practicum/shareit/user/UserRepository.java
@@ -3,9 +3,7 @@ package ru.practicum.shareit.user;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.transaction.annotation.Transactional;
 
-@Transactional
 public interface UserRepository extends JpaRepository<User, Long> {
 
     @Query("update User u set u.email = ?2, u.name = ?3 where u.id = ?1")
